Studies on the Hospital Anxiety and Depression Scale.

The data set includes 28 studies on 14 items measuring the Hospital Anxiety and Depression Scale (HADS) Reported by Norton et al. (2013) , with data and documentation copied over from the metaSEM package (Cheung 2015) .

Usage

Norton13

Format

Norton13

A list of data with the following structure:

data

A list of 28 studies of correlation matrices. The variables are 14 items (x1 to x14) measuring HADS

n

A vector of sample sizes

population

A vector of the population of the data

group

A vector of classification into patients vs. non-patients based on population

Source

https://cran.r-project.org/web/packages/metaSEM/index.html

References

Cheung MW (2015). “metaSEM: An R Package for Meta-Analysis using Structural Equation Modeling.” Frontiers in Psychology, 5(1521). doi:10.3389/fpsyg.2014.01521 .

Norton S, Cosco T, Doyle F, Done J, Sacker A (2013). “The Hospital Anxiety and Depression Scale: A meta confirmatory factor analysis.” Journal of Psychosomatic Research, 74(1), 74--81. ISSN 0022-3999, doi:10.1016/j.jpsychores.2012.10.010 .