Class "mbsem": For representing a model fitted with minorbsem

The mbsem class represents a fitted model. It contains summary statistics for structural parameters, summary statistics for standardized residual covariances, the data used to fit the model, and the fitted Stan model.

Objects from the Class

Objects can be created by calls to the minorbsem or meta_mbcfa function.

Slots

major_parameters:

Summary statistics for structural parameters

minor_factor_matrix:

Summary statistics for standardized residual covariances

data_list:

Data used to fit the model

priors:

Priors used to fit the model

stan_fit:

Fitted RStan or CmdStan model

version:

Package version used to fit model

Methods

logLik

signature(object = "mbsem", include_residuals = FALSE): Returns the casewise log-likelihood as long as the full data is available. include_residuals = FALSE ensures the minor factor residual covariances are not included when computing the log-likelihood. See casewise_log_likelihood for more details.

fitted

signature(object = "mbsem", include_residuals = TRUE): Returns the posterior distribution of the model-implied covariance matrix as a #(iterations) by #(items ^ 2) matrix. If include_residuals = TRUE, includes residual covariances induced by minor factors. If include_residuals = FALSE, excludes residual covariances induced by minor factors.

residuals

signature(object = "mbsem", standardized = TRUE): Returns the posterior distribution of residual covariances reflecting the influences of minor factors as a #(iterations) by #(items ^ 2) matrix. If standardized = TRUE, returns standardized residual covariances. If standardized = FALSE, returns UNstandardized residual covariances.

show

signature(object = "mbsem"): Pretty printing of model results. See pretty_print_summary for more printing options.

See also

Examples

  if (FALSE) {
    fit_1 <- minorbsem("F1 =~ x1 + x2 + x3
                        F2 =~ x4 + x5 + x6
                        F3 =~ x7 + x8 + x9", HS)
    # Model-implied covariances
    mod_imp_cov <- fitted(fit_1)
    # Get average of model-implied covariance matrix
    matrix(colMeans(mod_imp_cov), nrow = fit_1@data_list$Ni)
  }